linux-kthread: tid should be declared long
Otherwise the iterate_over_threads() iterator will never return
the correct GDB thread for a particular tid.

This overcomes the increasing gdb thread ID number ordering issue.
As when the kernel task list was re-enumerated the existing GDB
threads were not correctly matched with the process_t structs.

This had the side effect of new GDB threads being created everytime
via add_thread_with_info().

+++ gdb/linux-kthread.c
@@ -398,7 +398,7 @@ get_task_info (CORE_ADDR task_struct, process_t ** ps,
   size_t size;
   unsigned char *task_name;
   int i = 0;
-  int tid = 0;
+  long tid = 0;
   ptid_t this_ptid;
 
   while (*ps && (*ps)->valid)
